Update GUI plot parameters and log data file path
Updates were made to the plot parameters in the GUI including changes to the update_rate, max_points and plot_rate. The selection of different update rates was removed, providing a consistent plot behaviour. In addition, the path of the data file used is now logged, providing more comprehensive tracking information.